POC: Using sidecar logs to extract larger results
Prior to this, we were extracting results from tasks via the termination messages which had a limit of only 4 KB per pod.

We now run a dedicated sidecar that has access to the results of all the steps. This sidecar prints out the result and its content to stdout. The logs of the sidecar are parsed by the taskrun controller and the results updated instead of termination logs. This allows us a much higher capacity for storing results ~ 1.5 M (CRD limit).